+++ source/Examples/Resources/AdfsWebApiApplication/1-AdfsWebApiApplication_Config.ps1
@@ -19,7 +19,7 @@
 
 <#
     .DESCRIPTION
-        This configuration will add a Web API application role to an application in Active Directory Federation
+        This configuration will add a Web API application to an application group in Active Directory Federation
         Services (AD FS).
 #>
 
@@ -31,6 +31,12 @@ Configuration AdfsWebApiApplication_Config
 
     Node localhost
     {
+        AdfsApplicationGroup AppGroup1
+        {
+            Name        = 'AppGroup1'
+            Description = "This is the AppGroup1 Description"
+        }
+
         AdfsWebApiApplication WebApiApp1
         {
             Name                          = 'AppGroup1 - Web API'