+++ source/DSCResources/MSFT_AdfsApplicationPermission/MSFT_AdfsApplicationPermission.psm1
@@ -376,7 +376,8 @@ function Test-TargetResource
 
             $propertiesNotInDesiredState = (
                 Compare-ResourcePropertyState -CurrentValues $targetResource -DesiredValues $PSBoundParameters `
-                    @commonParms | Where-Object -Property InDesiredState -eq $false)
+                    -UnorderedArrayProperties 'ScopeNames' @commonParms | `
+                        Where-Object -Property InDesiredState -eq $false)
 
             if ($propertiesNotInDesiredState)
             {