+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - AdsfWebApiApplication
-  - Fixed issue when the `AllowedClientTypes` Property had multiple values ([#63](https://github.com/X-Guardian/AdfsDsc/issues/63)])
+  - Fixed issue when the `AllowedClientTypes` Property had multiple values ([#63](https://github.com/X-Guardian/AdfsDsc/issues/63)]).
+  - Fixed issue with comparing custom Issuance Transform Rules ([#57](https://github.com/X-Guardian/AdfsDsc/issues/57)).
 
 ## [1.3.0] - 2023-03-19