+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- AdfsDsc
+  - Fixed issue comparing arrays if elements were in a different order ([#67](https://github.com/X-Guardian/AdfsDsc/issues/67)]).
+
 ## [1.3.1] - 2023-04-19
 
 ### Fixed