+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - AdsfRelyingPartyTrust
   - Fixed issue when the `AllowedClientTypes` Property had multiple values.
 
+### Changed
+
+- AdfsDsc
+  - Updated the Help and Example files.
+
 ## [1.3.1] - 2023-04-19
 
 ### Fixed