+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - AdfsDsc
   - Fixed issue comparing arrays if elements were in a different order ([#67](https://github.com/X-Guardian/AdfsDsc/issues/67)]).
+- AdsfRelyingPartyTrust
+  - Fixed issue when the `AllowedClientTypes` Property had multiple values.
+
+### Changed
+
+- AdfsDsc
+  - Updated the Help and Example files.
 
 ## [1.3.1] - 2023-04-19